content/media/webaudio/test/test_AudioContext.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test whether we can create an AudioContext interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  ok(!window.AudioContext, "AudioContext should be hidden behind a pref");
  var accessThrows = false;
  try {
    new AudioContext();
  } catch (e) {
    accessThrows = true;
  }
  ok(accessThrows, "AudioContext should be hidden behind a pref");
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);
  var ac = new AudioContext();
  ok(ac, "Create a AudioContext object");
  is(ac.sampleRate, 44100, "Correct sample rate");
  SpecialPowers.clearUserPref("media.webaudio.enabled");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
